+++ packages/dash-components/src/components/dash-chip/dash-chip.css
@@ -66,8 +66,8 @@
   }
 
   &:focus-within {
-    --dash-focus-outline-max-offset: 1px;
-    --dash-focus-outline-offset: 0px;
+    --dash-focus-outline-offset-50: 1px;
+    --dash-focus-outline-offset-100: 0px;
     @extend .focus-outline;
   }
